backend/requirements.txt
asgiref==3.6.0
astroid==2.13.2
certifi==2022.12.7
cloudinary==1.31.0
colorama==0.4.6
dill==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==2.6.2
pylint==2.15.10
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
wrapt==1.14.1
